// packages/ui/src/lib/connect/viem/account-abstraction/types/userOperation.ts
export type EstimateUserOperationGasReturnType<
  entryPointVersion extends EntryPointVersion = EntryPointVersion,
  uint256 = bigint
> = OneOf<
  | (entryPointVersion extends "0.7"
      ? {
          preVerificationGas: uint256;
          verificationGasLimit: uint256;
          callGasLimit: uint256;
          paymasterVerificationGasLimit?: uint256 | undefined;
          paymasterPostOpGasLimit?: uint256 | undefined;
        }
      : never)
  | (entryPointVersion extends "0.6"
      ? {
          preVerificationGas: uint256;
          verificationGasLimit: uint256;
          callGasLimit: uint256;
        }
      : never)
>;